Refactored some MRO app code and moved helper codes to radp_library
